requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Supabase
supabase>=2.17.0

# OpenAI
openai>=1.12.0

# Authentication & Security
PyJWT>=2.10.1,<3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# HTTP & Networking
httpx>=0.27.0
requests>=2.31.0

# Environment management
python-dotenv>=1.0.1

# Image processing
Pillow>=10.4.0
pytesseract>=0.3.13
opencv-python>=4.12.0

# Video processing
yt-dlp>=2025.7.21
moviepy>=2.2.1

# Audio processing (for Whisper)
openai-whisper>=20231117

# Web scraping (for URL extraction)
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
python-slugify>=8.0.1
aiofiles>=23.0.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
